+++ tests/renderer/base/state/fixtures/appState.ts
@@ -63,7 +63,8 @@ const appState: AppState = {
     shelf: {
       widgetList: []
     },
-    worktable: {}
+    worktable: {},
+    widgetDynamicTitles: {}
   }
 };